tests/wpt/tests/css/css-borders/tentative/parsing/box-shadow-position-valid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Borders and Box Decorations 4 Test: Parsing 'box-shadow-position' with valid values</title>
<link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-borders-4/#box-shadow-position">
<meta name="assert" content="This test checks that 'box-shadow-position' supports positional keywords.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
test_valid_value("box-shadow-position", "inset");
test_valid_value("box-shadow-position", "outset");
test_valid_value("box-shadow-position", "inset, outset");
test_valid_value("box-shadow-position", "outset, inset");
</script>
